A branch to test angle of attack in unity
Also a test for testing tests in different branches. Having
the angle of attack test in a different branch allows us to
only store the files necessary to that test or scene. If we did everything
In one branch, the repo size would grow to be large, since
any objects we delete will stay in history.